Curve Editor: Change PixelsPerInput to double to fix jittery movement when manipulating keys
#rb mike.zyracki
#jira UE-184603

[CL 27886397 by max chen in ue5-main branch]

+++ Engine/Source/Editor/CurveEditor/Private/CurveEditor.cpp
@@ -690,7 +690,7 @@ void FCurveEditor::ZoomToFitInternal(EAxisList::Type Axes, const TMap<FCurveMode
 			TSharedPtr<SCurveEditorPanel> Panel = WeakPanel.Pin();
 			TSharedPtr<SCurveEditorView> View = WeakView.Pin();
 
-			int32 PanelWidth = 0;
+			double PanelWidth = 0;
 			if (Panel.IsValid())
 			{
 				PanelWidth = WeakPanel.Pin()->GetViewContainerGeometry().GetLocalSize().X;
@@ -700,7 +700,7 @@ void FCurveEditor::ZoomToFitInternal(EAxisList::Type Axes, const TMap<FCurveMode
 				PanelWidth = View->GetViewSpace().GetPhysicalWidth();
 			}
 			
-			double InputPercentage = PanelWidth != 0 ? FMath::Min(Settings->GetFrameInputPadding() / (float)PanelWidth, 0.5) : 0.1; // Cannot pad more than half the width
+			double InputPercentage = PanelWidth != 0 ? FMath::Min(Settings->GetFrameInputPadding() / PanelWidth, 0.5) : 0.1; // Cannot pad more than half the width
 
 			const double MinInputZoom = InputSnapEnabledAttribute.Get() ? InputSnapRateAttribute.Get().AsInterval() : 0.00001;
 			const double InputPadding = FMath::Max((InputMax - InputMin) * InputPercentage, MinInputZoom);
@@ -732,7 +732,7 @@ void FCurveEditor::ZoomToFitInternal(EAxisList::Type Axes, const TMap<FCurveMode
 		{
 			TSharedPtr<SCurveEditorPanel> Panel = WeakPanel.Pin();
 
-			int32 PanelHeight = 0;
+			double PanelHeight = 0;
 			if (Panel.IsValid())
 			{
 				PanelHeight = WeakPanel.Pin()->GetViewContainerGeometry().GetLocalSize().Y;
@@ -742,7 +742,7 @@ void FCurveEditor::ZoomToFitInternal(EAxisList::Type Axes, const TMap<FCurveMode
 				PanelHeight = View->GetViewSpace().GetPhysicalHeight();
 			}
 
-			double OutputPercentage = PanelHeight != 0 ? FMath::Min(Settings->GetFrameOutputPadding() / (float)PanelHeight, 0.5) : 0.1; // Cannot pad more than half the height
+			double OutputPercentage = PanelHeight != 0 ? FMath::Min(Settings->GetFrameOutputPadding() / PanelHeight, 0.5) : 0.1; // Cannot pad more than half the height
 
 			constexpr double MinOutputZoom = 0.00001;
 			const double OutputPadding = FMath::Max((OutputMax - OutputMin) * OutputPercentage, MinOutputZoom);
